refactor(payments): move payment types to core
