+++ packages/blockchain-wallet-v4-frontend/src/data/components/borrow/sagas.utils.ts
@@ -3,11 +3,10 @@ import * as S from './selectors'
 import { ADDRESS_TYPES } from 'blockchain-wallet-v4/src/redux/payment/btc/utils'
 import { APIType } from 'core/network/api'
 import { call, CallEffect, put, select } from 'redux-saga/effects'
-import { CoinType, RemoteDataType } from 'core/types'
+import { CoinType, PaymentType, PaymentValue, RemoteDataType } from 'core/types'
 import { convertStandardToBase } from '../exchange/services'
 import { Exchange } from 'blockchain-wallet-v4/src'
 import { NO_OFFER_EXISTS } from './model'
-import { PaymentType, PaymentValue } from './types'
 import { promptForSecondPassword } from 'services/SagaService'
 import BigNumber from 'bignumber.js'
