+++ packages/blockchain-wallet-v4-frontend/src/modals/Borrow/BorrowForm/index.tsx
@@ -1,9 +1,10 @@
 import { actions } from 'data'
 import { bindActionCreators, compose, Dispatch } from 'redux'
-import { BorrowMinMaxType, PaymentType, RatesType } from 'data/types'
+import { BorrowMinMaxType, RatesType } from 'data/types'
 import {
   CoinType,
   OfferType,
+  PaymentType,
   RemoteDataType,
   SupportedCoinsType
 } from 'core/types'
