+++ packages/blockchain-wallet-v4-frontend/src/modals/Borrow/ConfirmBorrow/index.tsx
@@ -2,8 +2,8 @@ import { actions } from 'data'
 import { bindActionCreators, compose, Dispatch } from 'redux'
 import { connect } from 'react-redux'
 import { getData } from './selectors'
-import { OfferType, RemoteDataType } from 'core/types'
-import { PaymentValue, RatesType } from 'data/types'
+import { OfferType, PaymentValue, RemoteDataType } from 'core/types'
+import { RatesType } from 'data/types'
 import { RootState } from 'data/rootReducer'
 import DataError from 'components/DataError'
 import Loading from './template.loading'
