+++ packages/blockchain-wallet-v4/src/types/index.ts
@@ -53,6 +53,7 @@ export type BtcAccountType = {
 
 export type AccountTypes = BtcAccountType
 
+export * from '../redux/payment/types'
 export * from '../redux/walletOptions/types'
 export * from '../network/api/borrow/types'
 export * from '../network/types'
